solo falta la parte de mecott

+++ vercursos.php
@@ -51,7 +51,7 @@ require "conexion.php";
                         echo "<td>{$row['nombre_docente']}</td>";
                         echo "<td>{$row['numero_horas']}</td>";
                         echo "<td>{$row['dias']}</td>";
-                        echo "<td><a href='eliminar_curso.php?id={$row['id']}'>Eliminar</a></td>";
+                        echo "<td><a href='eliminar_curso.php?id={$row['id']}' onclick='confirmarEliminacion(event)'>Eliminar</a></td>";
                         echo "</tr>";
                     }
                 } else {
@@ -64,6 +64,13 @@ require "conexion.php";
         </div>
     </div>
 </div>
+<script>
+function confirmarEliminacion(event) {
+    if (!confirm("¿Seguro que quieres eliminar este registro?")) {
+        event.preventDefault(); 
+    }
+}
+</script>
 
 </body>
 </html>
